feat: move images from connector-runtime-docker repository
