+++ .github/workflows/DEPLOY.yaml
@@ -83,6 +83,15 @@ jobs:
           username: ${{ steps.secrets.outputs.DOCKERHUB_USER }}
           password: ${{ steps.secrets.outputs.DOCKERHUB_PASSWORD }}
 
+      - name: Build and Push Docker Image tag SNAPSHOT - connector-runtime
+        uses: docker/build-push-action@v4
+        with:
+          context: connector-runtime/connector-runtime-application/
+          push: ${{ github.ref == 'refs/heads/main' }}
+          tags: camunda/connectors:SNAPSHOT
+          platforms: linux/amd64,linux/arm64
+          provenance: false
+
       - name: Build and Push Docker Image tag SNAPSHOT - bundle-default
         uses: docker/build-push-action@v4
         with:
