+++ bundle/default-bundle/Dockerfile
@@ -1,10 +1,10 @@
 FROM eclipse-temurin:17.0.7_7-jre
 
 RUN mkdir /opt/app
-# Download connectors from maven central
 COPY target/*-with-dependencies.jar /opt/app/
 
-COPY start.sh /start.sh
+# Using the start script from the base connector runtime image
+COPY ../../connector-runtime/connector-runtime-application/start.sh /start.sh
 RUN chmod +x start.sh
 
 ENTRYPOINT ["/start.sh"]
