+++ bundle/default-bundle/pom.xml
@@ -13,7 +13,7 @@
   <dependencies>
     <dependency>
       <groupId>io.camunda.connector</groupId>
-      <artifactId>spring-boot-starter-camunda-connectors</artifactId>
+      <artifactId>connector-runtime-application</artifactId>
     </dependency>
     <dependency>
       <groupId>io.micrometer</groupId>
@@ -88,7 +88,7 @@
         <groupId>org.codehaus.mojo</groupId>
         <artifactId>exec-maven-plugin</artifactId>
         <configuration>
-          <mainClass>io.camunda.connector.bundle.ConnectorRuntimeApplication</mainClass>
+          <mainClass>io.camunda.connector.runtime.app.ConnectorRuntimeApplication</mainClass>
         </configuration>
       </plugin>
        <plugin>
@@ -143,7 +143,7 @@
                   <resource>META-INF/io.netty.versions.properties</resource>
               </transformer>
               <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                <mainClass>io.camunda.connector.bundle.ConnectorRuntimeApplication</mainClass>
+                <mainClass>io.camunda.connector.runtime.app.ConnectorRuntimeApplication</mainClass>
               </transformer>
             </transformers>
           </configuration>
