--- bundle/mvn/default-bundle/src/test/resources/application.properties
@@ -1,19 +1,0 @@
-# Configuration for running connectors locally in bundle with connector-runtime
-server.port=9898
-#zeebe.broker.gateway-address=localhost:26500
-#zeebe.client.security.plaintext=true
-
-management.server.port=9080
-management.context-path=/actuator
-management.endpoints.web.exposure.include=metrics,health,prometheus
-
-camunda.connector.polling.enabled=true
-camunda.connector.polling.interval=5000
-
-camunda.operate.client.url=http://localhost:8081
-camunda.operate.client.username=demo
-camunda.operate.client.password=demo
-
-logging.level.io.camunda.connectors=TRACE
-
-operate.client.enabled=true
