+++ bundle/pom.xml
@@ -31,6 +31,11 @@
       </dependency>
       <dependency>
         <groupId>io.camunda.connector</groupId>
+        <artifactId>connector-runtime-application</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>io.camunda.connector</groupId>
         <artifactId>connector-aws-sqs</artifactId>
         <version>${project.version}</version>
       </dependency>
